Modify index of blogs controller
The index action on the blogs controller now checks for if a user is
currently logged in.  If so, it gets all the blogs that match their
token to owner of the blogs.  If there is no token (or the token
doesn't match) then it gets all of the blogs regardless of the
owner.  This makes it so people who are not signed in (guests) can see
the content of other's blogs when they visit the page, but users who
are logged in will see only the blogs they have created.

Next step:  start the pages back-end

Contributors: J.A.M.S.